+++ src/gui/ComboUnitsAngle.cpp
@@ -29,10 +29,10 @@
 ComboUnitsAngle::ComboUnitsAngle( QWidget *parent ) :
     ComboUnits ( parent )
 {
-    _coefs.push_back( 1.0f );
+    _coefs.push_back( 1.0 );
     _names.push_back( QString( "rad" ) );
 
-    _coefs.push_back( (float)fdm::Units::rad2deg() );
+    _coefs.push_back( fdm::Units::rad2deg() );
     _names.push_back( QString( "deg" ) );
 
     for ( size_t i = 0; i < _names.size(); i++ ) addItem( _names[ i ] );
