+++ src/gui/ComboUnitsFlowRate.cpp
@@ -29,16 +29,16 @@
 ComboUnitsFlowRate::ComboUnitsFlowRate( QWidget *parent ) :
     ComboUnits ( parent )
 {
-    _coefs.push_back( 1.0f );
+    _coefs.push_back( 1.0 );
     _names.push_back( QString( "kg/s" ) );
 
-    _coefs.push_back( 3600.0f );
+    _coefs.push_back( 3600.0 );
     _names.push_back( QString( "kg/h" ) );
 
-    _coefs.push_back( (float)fdm::Units::kg2lb() );
+    _coefs.push_back( fdm::Units::kg2lb() );
     _names.push_back( QString( "lb/s" ) );
 
-    _coefs.push_back( 3600.0f * (float)fdm::Units::kg2lb() );
+    _coefs.push_back( 3600.0 * fdm::Units::kg2lb() );
     _names.push_back( QString( "lb/h" ) );
 
     for ( size_t i = 0; i < _names.size(); i++ ) addItem( _names[ i ] );
