+++ src/gui/ComboUnitsLength.cpp
@@ -29,19 +29,19 @@
 ComboUnitsLength::ComboUnitsLength( QWidget *parent ) :
     ComboUnits ( parent )
 {
-    _coefs.push_back( 1.0f );
+    _coefs.push_back( 1.0 );
     _names.push_back( QString( "m" ) );
 
-    _coefs.push_back( (float)fdm::Units::m2ft() );
+    _coefs.push_back( fdm::Units::m2ft() );
     _names.push_back( QString( "ft" ) );
 
-    _coefs.push_back( (float)fdm::Units::m2km() );
+    _coefs.push_back( fdm::Units::m2km() );
     _names.push_back( QString( "km" ) );
 
-    _coefs.push_back( (float)fdm::Units::m2mi() );
+    _coefs.push_back( fdm::Units::m2mi() );
     _names.push_back( QString( "mi" ) );
 
-    _coefs.push_back( (float)fdm::Units::m2nmi() );
+    _coefs.push_back( fdm::Units::m2nmi() );
     _names.push_back( QString( "nmi" ) );
 
     for ( size_t i = 0; i < _names.size(); i++ ) addItem( _names[ i ] );
