+++ src/gui/ComboUnitsRotation.cpp
@@ -29,13 +29,13 @@
 ComboUnitsRotation::ComboUnitsRotation( QWidget *parent ) :
     ComboUnits ( parent )
 {
-    _coefs.push_back( 1.0f );
+    _coefs.push_back( 1.0 );
     _names.push_back( QString( "rad/s" ) );
 
-    _coefs.push_back( (float)fdm::Units::rad2deg() );
+    _coefs.push_back( fdm::Units::rad2deg() );
     _names.push_back( QString( "deg/s" ) );
 
-    _coefs.push_back( (float)( 60.0f / ( 2.0f * M_PI ) ) );
+    _coefs.push_back( 60.0 / ( 2.0 * M_PI ) );
     _names.push_back( QString( "rpm" ) );
 
     for ( size_t i = 0; i < _names.size(); i++ ) addItem( _names[ i ] );
