+++ src/gui/ComboUnitsTemperature.cpp
@@ -29,13 +29,13 @@
 ComboUnitsTemperature::ComboUnitsTemperature( QWidget *parent ) :
     ComboUnits ( parent )
 {
-    _coefs.push_back( 1.0f );
+    _coefs.push_back( 1.0 );
     _names.push_back( QString( "K" ) );
 
-    _coefs.push_back( 1.0f );
+    _coefs.push_back( 1.0 );
     _names.push_back( QString( "deg C" ) );
 
-    _coefs.push_back( 1.0f );
+    _coefs.push_back( 1.0 );
     _names.push_back( QString( "deg F" ) );
 
     for ( size_t i = 0; i < _names.size(); i++ ) addItem( _names[ i ] );
@@ -47,9 +47,9 @@ ComboUnitsTemperature::~ComboUnitsTemperature() {}
 
 ////////////////////////////////////////////////////////////////////////////////
 
-float ComboUnitsTemperature::convert( float value ) const
+double ComboUnitsTemperature::convert( double value ) const
 {
-    float result = value;
+    double result = value;
 
     switch ( _index )
     {
@@ -62,9 +62,9 @@ float ComboUnitsTemperature::convert( float value ) const
 
 ////////////////////////////////////////////////////////////////////////////////
 
-float ComboUnitsTemperature::convertPrev( float value ) const
+double ComboUnitsTemperature::convertPrev( double value ) const
 {
-    float result = value;
+    double result = value;
 
     switch ( _index_prev )
     {
@@ -77,9 +77,9 @@ float ComboUnitsTemperature::convertPrev( float value ) const
 
 ////////////////////////////////////////////////////////////////////////////////
 
-float ComboUnitsTemperature::invert( float value ) const
+double ComboUnitsTemperature::invert( double value ) const
 {
-    float result = value;
+    double result = value;
 
     switch ( _index )
     {
@@ -92,9 +92,9 @@ float ComboUnitsTemperature::invert( float value ) const
 
 ////////////////////////////////////////////////////////////////////////////////
 
-float ComboUnitsTemperature::invertPrev( float value ) const
+double ComboUnitsTemperature::invertPrev( double value ) const
 {
-    float result = value;
+    double result = value;
 
     switch ( _index_prev )
     {
